Sample mining.notify for block 100,000 is pretty much done
Add explainer on hex2bin_tbl - smart
